Remove install requirement
With the right structure, it is not necessary to install the project before
running tests.  It is probably(?) not good practice to have to do this.

Instead, if the tests are in a directory 'tests' and they properly include
modules & submodules in a directory <module_name> then running
`python -m pytest` from the repo root directory should work.